Added SCW node timeout, request limit/backoff
