+++ discord/message_handlers.go
@@ -15,7 +15,7 @@ import (
 	"github.com/nicksnyder/go-i18n/v2/i18n"
 )
 
-const RateLimitNodeThreshold = 8000
+var RateLimitNodeThreshold = 8000
 
 const downloadURL = "https://github.com/denverquane/amonguscapture/releases/latest/download/amonguscapture.exe"
